refactor: create new vo directory
