+++ BE/src/main/java/com/codesquad/issue04/domain/issue/vo/Emoji.java
@@ -1,4 +1,4 @@
-package com.codesquad.issue04.domain.issue;
+package com.codesquad.issue04.domain.issue.vo;
 
 import java.io.Serializable;
