+++ BE/src/main/java/com/codesquad/issue04/domain/issue/vo/Photo.java
@@ -1,4 +1,4 @@
-package com.codesquad.issue04.domain.issue;
+package com.codesquad.issue04.domain.issue.vo;
 
 import javax.persistence.Embeddable;
